sql/index: remove pilosa driver and remove pilosalib to pilosa
