+++ .gitignore
@@ -30,3 +30,5 @@ Makefile.main
 .ci/
 _example/main
 _example/*.exe
+
+go.sum
