+++ sql/index/pilosa/index.go
@@ -4,16 +4,61 @@ import (
 	"context"
 	"sync"
 
+	"github.com/pilosa/pilosa"
 	errors "gopkg.in/src-d/go-errors.v1"
-
-	pilosa "github.com/pilosa/go-pilosa"
 	"gopkg.in/src-d/go-mysql-server.v0/sql"
 	"gopkg.in/src-d/go-mysql-server.v0/sql/index"
 )
 
+// concurrentPilosaIndex is a wrapper of pilosa.Index that can be opened and closed
+// concurrently.
+type concurrentPilosaIndex struct {
+	*pilosa.Index
+	m  sync.Mutex
+	rc int
+}
+
+func newConcurrentPilosaIndex(idx *pilosa.Index) *concurrentPilosaIndex {
+	return &concurrentPilosaIndex{Index: idx}
+}
+
+func (i *concurrentPilosaIndex) Open() error {
+	i.m.Lock()
+	defer i.m.Unlock()
+
+	if i.rc == 0 {
+		if err := i.Index.Open(); err != nil {
+			return err
+		}
+	}
+
+	i.rc++
+	return nil
+}
+
+func (i *concurrentPilosaIndex) Close() error {
+	i.m.Lock()
+	defer i.m.Unlock()
+
+	i.rc--
+	if i.rc < 0 {
+		i.rc = 0
+	}
+
+	if i.rc == 0 {
+		return i.Index.Close()
+	}
+
+	return nil
+}
+
+var (
+	errInvalidKeys = errors.NewKind("expecting %d keys for index %q, got %d")
+)
+
 // pilosaIndex is an pilosa implementation of sql.Index interface
 type pilosaIndex struct {
-	client  *pilosa.Client
+	index   *concurrentPilosaIndex
 	mapping *mapping
 	cancel  context.CancelFunc
 	wg      sync.WaitGroup
@@ -24,21 +69,16 @@ type pilosaIndex struct {
 	expressions []string
 }
 
-func newPilosaIndex(mappingfile string, client *pilosa.Client, cfg *index.Config) *pilosaIndex {
+func newPilosaIndex(idx *pilosa.Index, mapping *mapping, cfg *index.Config) *pilosaIndex {
 	return &pilosaIndex{
-		client:      client,
+		index:       newConcurrentPilosaIndex(idx),
 		db:          cfg.DB,
 		table:       cfg.Table,
 		id:          cfg.ID,
 		expressions: cfg.Expressions,
-		mapping:     newMapping(mappingfile),
+		mapping:     mapping,
 	}
 }
-
-var (
-	errInvalidKeys = errors.NewKind("expecting %d keys for index %q, got %d")
-	errPilosaQuery = errors.NewKind("error executing pilosa query: %s")
-)
 
 // Get returns an IndexLookup for the given key in the index.
 // If key parameter is not present then the returned iterator
@@ -48,20 +88,9 @@ func (idx *pilosaIndex) Get(keys ...interface{}) (sql.IndexLookup, error) {
 		return nil, errInvalidKeys.New(len(idx.expressions), idx.ID(), len(keys))
 	}
 
-	schema, err := idx.client.Schema()
-	if err != nil {
-		return nil, err
-	}
-
-	index := schema.Index(indexName(idx.Database(), idx.Table()))
-	if err := idx.client.EnsureIndex(index); err != nil {
-		return nil, err
-	}
-
 	return &indexLookup{
 		id:          idx.ID(),
-		index:       index,
-		client:      idx.client,
+		index:       idx.index,
 		mapping:     idx.mapping,
 		keys:        keys,
 		expressions: idx.expressions,
@@ -70,7 +99,9 @@ func (idx *pilosaIndex) Get(keys ...interface{}) (sql.IndexLookup, error) {
 
 // Has checks if the given key is present in the index mapping
 func (idx *pilosaIndex) Has(p sql.Partition, key ...interface{}) (bool, error) {
-	idx.mapping.open()
+	if err := idx.mapping.open(); err != nil {
+		return false, err
+	}
 	defer idx.mapping.close()
 
 	n := len(key)
@@ -78,8 +109,6 @@ func (idx *pilosaIndex) Has(p sql.Partition, key ...interface{}) (bool, error) {
 		n = len(idx.expressions)
 	}
 
-	// We can make this loop parallel, but does it make sense?
-	// For how many (maximum) keys will be asked by one function call?
 	for i, expr := range idx.expressions {
 		name := fieldName(idx.ID(), expr, p)
 
@@ -121,31 +150,13 @@ func (idx *pilosaIndex) AscendGreaterOrEqual(keys ...interface{}) (sql.IndexLook
 		return nil, errInvalidKeys.New(len(idx.expressions), idx.ID(), len(keys))
 	}
 
-	schema, err := idx.client.Schema()
-	if err != nil {
-		return nil, err
-	}
-
-	index := schema.Index(indexName(idx.Database(), idx.Table()))
-	if err := idx.client.EnsureIndex(index); err != nil {
-		return nil, err
-	}
-
-	l := &ascendLookup{
-		filteredLookup: &filteredLookup{
-			id:          idx.ID(),
-			client:      idx.client,
-			index:       index,
-			mapping:     idx.mapping,
-			keys:        keys,
-			expressions: idx.expressions,
-		},
-		gte: keys,
-		lt:  nil,
-	}
-	l.initFilter()
-
-	return l, nil
+	return newAscendLookup(&filteredLookup{
+		id:          idx.ID(),
+		index:       idx.index,
+		mapping:     idx.mapping,
+		keys:        keys,
+		expressions: idx.expressions,
+	}, keys, nil), nil
 }
 
 func (idx *pilosaIndex) AscendLessThan(keys ...interface{}) (sql.IndexLookup, error) {
@@ -153,31 +164,13 @@ func (idx *pilosaIndex) AscendLessThan(keys ...interface{}) (sql.IndexLookup, er
 		return nil, errInvalidKeys.New(len(idx.expressions), idx.ID(), len(keys))
 	}
 
-	schema, err := idx.client.Schema()
-	if err != nil {
-		return nil, err
-	}
-
-	index := schema.Index(indexName(idx.Database(), idx.Table()))
-	if err := idx.client.EnsureIndex(index); err != nil {
-		return nil, err
-	}
-
-	l := &ascendLookup{
-		filteredLookup: &filteredLookup{
-			id:          idx.ID(),
-			client:      idx.client,
-			index:       index,
-			mapping:     idx.mapping,
-			keys:        keys,
-			expressions: idx.expressions,
-		},
-		gte: nil,
-		lt:  keys,
-	}
-	l.initFilter()
-
-	return l, nil
+	return newAscendLookup(&filteredLookup{
+		id:          idx.ID(),
+		index:       idx.index,
+		mapping:     idx.mapping,
+		keys:        keys,
+		expressions: idx.expressions,
+	}, nil, keys), nil
 }
 
 func (idx *pilosaIndex) AscendRange(greaterOrEqual, lessThan []interface{}) (sql.IndexLookup, error) {
@@ -189,30 +182,12 @@ func (idx *pilosaIndex) AscendRange(greaterOrEqual, lessThan []interface{}) (sql
 		return nil, errInvalidKeys.New(len(idx.expressions), idx.ID(), len(lessThan))
 	}
 
-	schema, err := idx.client.Schema()
-	if err != nil {
-		return nil, err
-	}
-
-	index := schema.Index(indexName(idx.Database(), idx.Table()))
-	if err := idx.client.EnsureIndex(index); err != nil {
-		return nil, err
-	}
-
-	l := &ascendLookup{
-		filteredLookup: &filteredLookup{
-			id:          idx.ID(),
-			client:      idx.client,
-			index:       index,
-			mapping:     idx.mapping,
-			expressions: idx.expressions,
-		},
-		gte: greaterOrEqual,
-		lt:  lessThan,
-	}
-	l.initFilter()
-
-	return l, nil
+	return newAscendLookup(&filteredLookup{
+		id:          idx.ID(),
+		index:       idx.index,
+		mapping:     idx.mapping,
+		expressions: idx.expressions,
+	}, greaterOrEqual, lessThan), nil
 }
 
 func (idx *pilosaIndex) DescendGreater(keys ...interface{}) (sql.IndexLookup, error) {
@@ -220,32 +195,14 @@ func (idx *pilosaIndex) DescendGreater(keys ...interface{}) (sql.IndexLookup, er
 		return nil, errInvalidKeys.New(len(idx.expressions), idx.ID(), len(keys))
 	}
 
-	schema, err := idx.client.Schema()
-	if err != nil {
-		return nil, err
-	}
-
-	index := schema.Index(indexName(idx.Database(), idx.Table()))
-	if err := idx.client.EnsureIndex(index); err != nil {
-		return nil, err
-	}
-
-	l := &descendLookup{
-		filteredLookup: &filteredLookup{
-			id:          idx.ID(),
-			client:      idx.client,
-			index:       index,
-			mapping:     idx.mapping,
-			keys:        keys,
-			expressions: idx.expressions,
-			reverse:     true,
-		},
-		gt:  keys,
-		lte: nil,
-	}
-	l.initFilter()
-
-	return l, nil
+	return newDescendLookup(&filteredLookup{
+		id:          idx.ID(),
+		index:       idx.index,
+		mapping:     idx.mapping,
+		keys:        keys,
+		expressions: idx.expressions,
+		reverse:     true,
+	}, keys, nil), nil
 }
 
 func (idx *pilosaIndex) DescendLessOrEqual(keys ...interface{}) (sql.IndexLookup, error) {
@@ -253,32 +210,14 @@ func (idx *pilosaIndex) DescendLessOrEqual(keys ...interface{}) (sql.IndexLookup
 		return nil, errInvalidKeys.New(len(idx.expressions), idx.ID(), len(keys))
 	}
 
-	schema, err := idx.client.Schema()
-	if err != nil {
-		return nil, err
-	}
-
-	index := schema.Index(indexName(idx.Database(), idx.Table()))
-	if err := idx.client.EnsureIndex(index); err != nil {
-		return nil, err
-	}
-
-	l := &descendLookup{
-		filteredLookup: &filteredLookup{
-			id:          idx.ID(),
-			client:      idx.client,
-			index:       index,
-			mapping:     idx.mapping,
-			keys:        keys,
-			expressions: idx.expressions,
-			reverse:     true,
-		},
-		gt:  nil,
-		lte: keys,
-	}
-	l.initFilter()
-
-	return l, nil
+	return newDescendLookup(&filteredLookup{
+		id:          idx.ID(),
+		index:       idx.index,
+		mapping:     idx.mapping,
+		keys:        keys,
+		expressions: idx.expressions,
+		reverse:     true,
+	}, nil, keys), nil
 }
 
 func (idx *pilosaIndex) DescendRange(lessOrEqual, greaterThan []interface{}) (sql.IndexLookup, error) {
@@ -290,31 +229,13 @@ func (idx *pilosaIndex) DescendRange(lessOrEqual, greaterThan []interface{}) (sq
 		return nil, errInvalidKeys.New(len(idx.expressions), idx.ID(), len(greaterThan))
 	}
 
-	schema, err := idx.client.Schema()
-	if err != nil {
-		return nil, err
-	}
-
-	index := schema.Index(indexName(idx.Database(), idx.Table()))
-	if err := idx.client.EnsureIndex(index); err != nil {
-		return nil, err
-	}
-
-	l := &descendLookup{
-		filteredLookup: &filteredLookup{
-			id:          idx.ID(),
-			client:      idx.client,
-			index:       index,
-			mapping:     idx.mapping,
-			expressions: idx.expressions,
-			reverse:     true,
-		},
-		gt:  greaterThan,
-		lte: lessOrEqual,
-	}
-	l.initFilter()
-
-	return l, nil
+	return newDescendLookup(&filteredLookup{
+		id:          idx.ID(),
+		index:       idx.index,
+		mapping:     idx.mapping,
+		expressions: idx.expressions,
+		reverse:     true,
+	}, greaterThan, lessOrEqual), nil
 }
 
 func (idx *pilosaIndex) Not(keys ...interface{}) (sql.IndexLookup, error) {
@@ -322,22 +243,103 @@ func (idx *pilosaIndex) Not(keys ...interface{}) (sql.IndexLookup, error) {
 		return nil, errInvalidKeys.New(len(idx.expressions), idx.ID(), len(keys))
 	}
 
-	schema, err := idx.client.Schema()
-	if err != nil {
-		return nil, err
-	}
-
-	index := schema.Index(indexName(idx.Database(), idx.Table()))
-	if err := idx.client.EnsureIndex(index); err != nil {
-		return nil, err
-	}
-
 	return &negateLookup{
 		id:          idx.ID(),
-		client:      idx.client,
-		index:       index,
+		index:       idx.index,
 		mapping:     idx.mapping,
 		keys:        keys,
 		expressions: idx.expressions,
 	}, nil
+}
+
+func newAscendLookup(f *filteredLookup, gte []interface{}, lt []interface{}) *ascendLookup {
+	l := &ascendLookup{filteredLookup: f, gte: gte, lt: lt}
+	if l.filter == nil {
+		l.filter = func(i int, value []byte) (bool, error) {
+			var v interface{}
+			var err error
+			if len(l.gte) > 0 {
+				v, err = decodeGob(value, l.gte[i])
+				if err != nil {
+					return false, err
+				}
+
+				cmp, err := compare(v, l.gte[i])
+				if err != nil {
+					return false, err
+				}
+
+				if cmp < 0 {
+					return false, nil
+				}
+			}
+
+			if len(l.lt) > 0 {
+				if v == nil {
+					v, err = decodeGob(value, l.lt[i])
+					if err != nil {
+						return false, err
+					}
+				}
+
+				cmp, err := compare(v, l.lt[i])
+				if err != nil {
+					return false, err
+				}
+
+				if cmp >= 0 {
+					return false, nil
+				}
+			}
+
+			return true, nil
+		}
+	}
+	return l
+}
+
+func newDescendLookup(f *filteredLookup, gt []interface{}, lte []interface{}) *descendLookup {
+	l := &descendLookup{filteredLookup: f, gt: gt, lte: lte}
+	if l.filter == nil {
+		l.filter = func(i int, value []byte) (bool, error) {
+			var v interface{}
+			var err error
+			if len(l.gt) > 0 {
+				v, err = decodeGob(value, l.gt[i])
+				if err != nil {
+					return false, err
+				}
+
+				cmp, err := compare(v, l.gt[i])
+				if err != nil {
+					return false, err
+				}
+
+				if cmp <= 0 {
+					return false, nil
+				}
+			}
+
+			if len(l.lte) > 0 {
+				if v == nil {
+					v, err = decodeGob(value, l.lte[i])
+					if err != nil {
+						return false, err
+					}
+				}
+
+				cmp, err := compare(v, l.lte[i])
+				if err != nil {
+					return false, err
+				}
+
+				if cmp > 0 {
+					return false, nil
+				}
+			}
+
+			return true, nil
+		}
+	}
+	return l
 }
