+++ sql/index/pilosa/mapping.go
@@ -15,7 +15,7 @@ import (
 // mapping
 // buckets:
 // - index name: columndID uint64 -> location []byte
-// - frame name: value []byte (gob encoding) -> rowID uint64
+// - field name: value []byte (gob encoding) -> rowID uint64
 type mapping struct {
 	path string
 
@@ -35,21 +35,35 @@ func newMapping(path string) *mapping {
 	return &mapping{path: path}
 }
 
-func (m *mapping) open() {
-	m.openCreate(false)
+func (m *mapping) open() error {
+	return m.openCreate(false)
 }
 
 // openCreate opens and sets creation mode in the database.
-func (m *mapping) openCreate(create bool) {
+func (m *mapping) openCreate(create bool) error {
 	m.clientMut.Lock()
 	defer m.clientMut.Unlock()
+	m.mut.Lock()
+	defer m.mut.Unlock()
+
+	if m.clients == 0 && m.db == nil {
+		var err error
+		m.db, err = bolt.Open(m.path, 0640, nil)
+		if err != nil {
+			return err
+		}
+	}
+
 	m.clients++
 	m.create = create
+	return nil
 }
 
 func (m *mapping) close() error {
 	m.clientMut.Lock()
 	defer m.clientMut.Unlock()
+	m.mut.Lock()
+	defer m.mut.Unlock()
 
 	if m.clients > 1 {
 		m.clients--
@@ -57,9 +71,6 @@ func (m *mapping) close() error {
 	}
 
 	m.clients = 0
-
-	m.mut.Lock()
-	defer m.mut.Unlock()
 
 	if m.db != nil {
 		if err := m.db.Close(); err != nil {
@@ -71,25 +82,8 @@ func (m *mapping) close() error {
 	return nil
 }
 
-func (m *mapping) query(fn func() error) error {
-	m.mut.Lock()
-	if m.db == nil {
-		var err error
-		m.db, err = bolt.Open(m.path, 0640, nil)
-		if err != nil {
-			m.mut.Unlock()
-			return err
-		}
-	}
-	m.mut.Unlock()
-
-	m.mut.RLock()
-	defer m.mut.RUnlock()
-	return fn()
-}
-
-func (m *mapping) rowID(frameName string, value interface{}) (uint64, error) {
-	val, err := m.get(frameName, value)
+func (m *mapping) rowID(fieldName string, value interface{}) (uint64, error) {
+	val, err := m.get(fieldName, value)
 	if err != nil {
 		return 0, err
 	}
@@ -169,100 +163,96 @@ func (m *mapping) transaction(writable bool, f func(*bolt.Tx) error) error {
 	return tx.Commit()
 }
 
-func (m *mapping) getRowID(frameName string, value interface{}) (uint64, error) {
+func (m *mapping) getRowID(fieldName string, value interface{}) (uint64, error) {
 	var id uint64
-	err := m.query(func() error {
-		var buf bytes.Buffer
-		enc := gob.NewEncoder(&buf)
-		err := enc.Encode(value)
+	var buf bytes.Buffer
+	enc := gob.NewEncoder(&buf)
+	err := enc.Encode(value)
+	if err != nil {
+		return 0, err
+	}
+
+	err = m.transaction(true, func(tx *bolt.Tx) error {
+		b, err := tx.CreateBucketIfNotExists([]byte(fieldName))
 		if err != nil {
 			return err
 		}
 
-		err = m.transaction(true, func(tx *bolt.Tx) error {
-			b, err := tx.CreateBucketIfNotExists([]byte(frameName))
-			if err != nil {
-				return err
-			}
+		key := buf.Bytes()
+		val := b.Get(key)
+		if val != nil {
+			id = binary.LittleEndian.Uint64(val)
+			return nil
+		}
 
-			key := buf.Bytes()
-			val := b.Get(key)
-			if val != nil {
-				id = binary.LittleEndian.Uint64(val)
-				return nil
-			}
+		// the first NextSequence is 1 so the first id will be 1
+		// this can only fail if the transaction is closed
+		id, _ = b.NextSequence()
 
-			// the first NextSequence is 1 so the first id will be 1
-			// this can only fail if the transaction is closed
-			id, _ = b.NextSequence()
-
-			val = make([]byte, 8)
-			binary.LittleEndian.PutUint64(val, id)
-			err = b.Put(key, val)
-			return err
-		})
-
+		val = make([]byte, 8)
+		binary.LittleEndian.PutUint64(val, id)
+		err = b.Put(key, val)
 		return err
 	})
+
+	if err != nil {
+		return 0, err
+	}
 
 	return id, err
 }
 
-func (m *mapping) getMaxRowID(frameName string) (uint64, error) {
+func (m *mapping) getMaxRowID(fieldName string) (uint64, error) {
 	var id uint64
-	err := m.query(func() error {
-		return m.transaction(true, func(tx *bolt.Tx) error {
-			b := tx.Bucket([]byte(frameName))
-			if b == nil {
-				return nil
-			}
-
-			id = b.Sequence()
+	err := m.transaction(true, func(tx *bolt.Tx) error {
+		b := tx.Bucket([]byte(fieldName))
+		if b == nil {
 			return nil
-		})
+		}
+
+		id = b.Sequence()
+		return nil
 	})
 
 	return id, err
 }
 
 func (m *mapping) putLocation(indexName string, colID uint64, location []byte) error {
-	return m.query(func() error {
-		return m.transaction(true, func(tx *bolt.Tx) error {
-			b, err := tx.CreateBucketIfNotExists([]byte(indexName))
-			if err != nil {
-				return err
-			}
+	return m.transaction(true, func(tx *bolt.Tx) error {
+		b, err := tx.CreateBucketIfNotExists([]byte(indexName))
+		if err != nil {
+			return err
+		}
 
-			key := make([]byte, 8)
-			binary.LittleEndian.PutUint64(key, colID)
+		key := make([]byte, 8)
+		binary.LittleEndian.PutUint64(key, colID)
 
-			return b.Put(key, location)
-		})
+		return b.Put(key, location)
 	})
 }
 
 func (m *mapping) sortedLocations(indexName string, cols []uint64, reverse bool) ([][]byte, error) {
 	var result [][]byte
-	err := m.query(func() error {
-		return m.db.View(func(tx *bolt.Tx) error {
-			b := tx.Bucket([]byte(indexName))
-			if b == nil {
-				return fmt.Errorf("bucket %s not found", indexName)
-			}
+	m.mut.RLock()
+	defer m.mut.RUnlock()
+	err := m.db.View(func(tx *bolt.Tx) error {
+		b := tx.Bucket([]byte(indexName))
+		if b == nil {
+			return fmt.Errorf("bucket %s not found", indexName)
+		}
 
-			for _, col := range cols {
-				key := make([]byte, 8)
-				binary.LittleEndian.PutUint64(key, col)
-				val := b.Get(key)
+		for _, col := range cols {
+			key := make([]byte, 8)
+			binary.LittleEndian.PutUint64(key, col)
+			val := b.Get(key)
 
-				// val will point to mmap addresses, so we need to copy the slice
-				dst := make([]byte, len(val))
-				copy(dst, val)
-				result = append(result, dst)
-			}
+			// val will point to mmap addresses, so we need to copy the slice
+			dst := make([]byte, len(val))
+			copy(dst, val)
+			result = append(result, dst)
+		}
 
-			return nil
-		})
+		return nil
 	})
 
 	if err != nil {
@@ -287,21 +277,17 @@ func (b byBytes) Less(i, j int) bool { return bytes.Compare(b[i], b[j]) < 0 }
 func (m *mapping) getLocation(indexName string, colID uint64) ([]byte, error) {
 	var location []byte
 
-	err := m.query(func() error {
-		err := m.transaction(true, func(tx *bolt.Tx) error {
-			b := tx.Bucket([]byte(indexName))
-			if b == nil {
-				return fmt.Errorf("bucket %s not found", indexName)
-			}
+	err := m.transaction(true, func(tx *bolt.Tx) error {
+		b := tx.Bucket([]byte(indexName))
+		if b == nil {
+			return fmt.Errorf("bucket %s not found", indexName)
+		}
 
-			key := make([]byte, 8)
-			binary.LittleEndian.PutUint64(key, colID)
+		key := make([]byte, 8)
+		binary.LittleEndian.PutUint64(key, colID)
 
-			location = b.Get(key)
-			return nil
-		})
-
-		return err
+		location = b.Get(key)
+		return nil
 	})
 
 	return location, err
@@ -311,17 +297,9 @@ func (m *mapping) getLocationFromBucket(
 	bucket *bolt.Bucket,
 	colID uint64,
 ) ([]byte, error) {
-	var location []byte
-
-	err := m.query(func() error {
-		key := make([]byte, 8)
-		binary.LittleEndian.PutUint64(key, colID)
-
-		location = bucket.Get(key)
-		return nil
-	})
-
-	return location, err
+	key := make([]byte, 8)
+	binary.LittleEndian.PutUint64(key, colID)
+	return bucket.Get(key), nil
 }
 
 func (m *mapping) getBucket(
@@ -330,20 +308,16 @@ func (m *mapping) getBucket(
 ) (*bolt.Bucket, error) {
 	var bucket *bolt.Bucket
 
-	err := m.query(func() error {
-		tx, err := m.db.Begin(writable)
-		if err != nil {
-			return err
-		}
+	tx, err := m.db.Begin(writable)
+	if err != nil {
+		return nil, err
+	}
 
-		bucket = tx.Bucket([]byte(indexName))
-		if bucket == nil {
-			tx.Rollback()
-			return fmt.Errorf("bucket %s not found", indexName)
-		}
-
-		return nil
-	})
+	bucket = tx.Bucket([]byte(indexName))
+	if bucket == nil {
+		_ = tx.Rollback()
+		return nil, fmt.Errorf("bucket %s not found", indexName)
+	}
 
 	return bucket, err
 }
@@ -351,49 +325,48 @@ func (m *mapping) getBucket(
 func (m *mapping) get(name string, key interface{}) ([]byte, error) {
 	var value []byte
 
-	err := m.query(func() error {
-		var buf bytes.Buffer
-		enc := gob.NewEncoder(&buf)
-		err := enc.Encode(key)
-		if err != nil {
-			return err
+	var buf bytes.Buffer
+	enc := gob.NewEncoder(&buf)
+	err := enc.Encode(key)
+	if err != nil {
+		return nil, err
+	}
+
+	err = m.transaction(true, func(tx *bolt.Tx) error {
+		b := tx.Bucket([]byte(name))
+		if b != nil {
+			value = b.Get(buf.Bytes())
+			return nil
 		}
 
-		return m.transaction(true, func(tx *bolt.Tx) error {
-			b := tx.Bucket([]byte(name))
-			if b != nil {
-				value = b.Get(buf.Bytes())
-				return nil
-			}
-
-			return fmt.Errorf("%s not found", name)
-		})
+		return fmt.Errorf("%s not found", name)
 	})
+
 	return value, err
 }
 
 func (m *mapping) filter(name string, fn func([]byte) (bool, error)) ([]uint64, error) {
 	var result []uint64
 
-	err := m.query(func() error {
-		return m.db.View(func(tx *bolt.Tx) error {
-			b := tx.Bucket([]byte(name))
-			if b == nil {
-				return nil
+	m.mut.RLock()
+	defer m.mut.RUnlock()
+	err := m.db.View(func(tx *bolt.Tx) error {
+		b := tx.Bucket([]byte(name))
+		if b == nil {
+			return nil
+		}
+
+		return b.ForEach(func(k, v []byte) error {
+			ok, err := fn(k)
+			if err != nil {
+				return err
 			}
 
-			return b.ForEach(func(k, v []byte) error {
-				ok, err := fn(k)
-				if err != nil {
-					return err
-				}
+			if ok {
+				result = append(result, binary.LittleEndian.Uint64(v))
+			}
 
-				if ok {
-					result = append(result, binary.LittleEndian.Uint64(v))
-				}
-
-				return nil
-			})
+			return nil
 		})
 	})
