+++ sql/index/pilosa/mapping_test.go
@@ -12,9 +12,8 @@ func TestRowID(t *testing.T) {
 	require := require.New(t)
 	setup(t)
 	defer cleanup(t)
-
 	m := newMapping(filepath.Join(tmpDir, "id.map"))
-	m.open()
+	require.NoError(m.open())
 	defer m.close()
 
 	cases := []int{0, 1, 2, 3, 4, 5, 5, 0, 3, 2, 1, 5}
@@ -37,7 +36,7 @@ func TestLocation(t *testing.T) {
 	defer cleanup(t)
 
 	m := newMapping(filepath.Join(tmpDir, "id.map"))
-	m.open()
+	require.NoError(m.open())
 	defer m.close()
 
 	cases := map[uint64]string{
@@ -66,7 +65,7 @@ func TestGet(t *testing.T) {
 	defer cleanup(t)
 
 	m := newMapping(filepath.Join(tmpDir, "id.map"))
-	m.open()
+	require.NoError(m.open())
 	defer m.close()
 
 	cases := []int{0, 1, 2, 3, 4, 5, 5, 0, 3, 2, 1, 5}
